Update and rename leetCode longest common prefix.py to Easy: Longest Common Prefix.py
Pointer 1 keeps a record of the length of the first element in the list.  Pointer 2 goes over each element in the list.  [x] of y ("y[x]" in the code) becomes "ffflllooiwwge", as it traverses every letter of every element in the list up to the "e" of "flower" due to the fact the next list element is "flow", but the pointer then stops as "flow" is the shortest list element.   If that result differs from strs[0][x] ("ffflllooowwweeerrr": each letter in the first list element times by the number of elements in the list), then stop adding them into the list named "result" and return the result. In this case there are only three "F"'s and three "L"'s, so they're returned.